requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
pyyaml>=6.0
pillow>=9.0.0
openai>=1.0.0
anthropic>=0.18.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
requests>=2.28.0
flask>=2.3.0
